requirements.txt
streamlit
requests
pydantic-ai
python-dotenv
airportsdata